+++ web/app/controllers/webhooks/shop_redact_controller.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-module ShopifyWebhooks
-  class CustomerRedactController < ApplicationController
+module Webhooks
+  class ShopRedactController < ApplicationController
     include ShopifyApp::WebhookVerification
 
     def receive
       webhook_request = ShopifyAPI::Webhooks::Request.new(raw_body: request.raw_post, headers: request.headers.to_h)
-      CustomerRedactJob.perform_later(shop_domain: webhook_request.shop, webhook: webhook_request.parsed_body)
+      ShopRedactJob.perform_later(shop_domain: webhook_request.shop, webhook: webhook_request.parsed_body)
       head(:no_content)
     end
   end
