+++ web/config/routes.rb
@@ -3,9 +3,6 @@
 Rails.application.routes.draw do
   root to: "home#index"
 
-  mount ShopifyApp::Engine, at: "/api"
-  get "/api", to: redirect(path: "/") # Needed because our engine root is /api but that breaks FE routing
-
   # If you are adding routes outside of the /api path, remember to also add a proxy rule for
   # them in web/frontend/vite.config.js
 
@@ -18,13 +15,16 @@ Rails.application.routes.draw do
         get :count
       end
     end
-    namespace :shopify_webhooks do
+    namespace :webhooks do
       post "/app_uninstalled", to: "app_uninstalled#receive"
       post "/customers_data_request", to: "customers_data_request#receive"
       post "/customers_redact", to: "customers_redact#receive"
       post "/shop_redact", to: "shop_redact#receive"
     end
   end
+
+  mount ShopifyApp::Engine, at: "/api"
+  get "/api", to: redirect(path: "/") # Needed because our engine root is /api but that breaks FE routing
 
   # Any other routes will just render the react app
   match "*path" => "home#index", via: [:get, :post]
